fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
crawl4ai==0.5.0.post8
fastapi==0.115.11
Flask==2.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.120.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-generativeai==0.8.4
google-oauth==1.0.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ollama==0.4.7
orjson==3.10.0
packaging==24.2
playwright==1.51.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pyparsing==3.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn[standard]==0.23.2
webdriver-manager==3.5.1
websockets==11.0.3
